refactor: add reuseable modal ModalWithExchangeTab
