+++ src/components/_tabs/ExchangeTab.tsx
@@ -11,24 +11,26 @@ export const ExchangeTab = () => {
 
   return (
     <Stack>
-      {exchanges.map((item) => (
-        <Link key={item.name} href={item.url} target="_blank">
-          <HStack
-            justifyContent="space-between"
-            backgroundColor="surface.secondary"
-            padding={4}
-            borderRadius="xl"
-          >
-            <HStack spacing={4}>
-              <Image alt={item.name} src={item.logo} boxSize={6} />
-              <Text fontSize="xl" fontWeight="bold">
-                {item.name}
-              </Text>
+      {exchanges
+        .filter((item) => item.name !== "Sommelier")
+        .map((item) => (
+          <Link key={item.name} href={item.url} target="_blank">
+            <HStack
+              justifyContent="space-between"
+              backgroundColor="surface.secondary"
+              padding={4}
+              borderRadius="xl"
+            >
+              <HStack spacing={4}>
+                <Image alt={item.name} src={item.logo} boxSize={6} />
+                <Text fontSize="xl" fontWeight="bold">
+                  {item.name}
+                </Text>
+              </HStack>
+              <ExternalLinkIcon />
             </HStack>
-            <ExternalLinkIcon />
-          </HStack>
-        </Link>
-      ))}
+          </Link>
+        ))}
     </Stack>
   )
 }
